requirements.txt
absl-py==1.4.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
filelock==3.10.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h5py==3.8.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
packaging @ file:///private/tmp/sip--packaging-20230420-6549-1c57q83/packaging-23.1
pandas==2.0.0
Pillow==9.4.0
ply==3.11
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQt6==6.5.0
PyQt6-3D==6.5.0
PyQt6-Charts==6.5.0
PyQt6-DataVisualization==6.5.0
PyQt6-NetworkAuth==6.5.0
PyQt6-sip==13.5.0
PyQt6-WebEngine==6.5.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
pytz==2023.3
QScintilla==2.14.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sip==6.7.9
six==1.16.0
soundfile==0.12.1
sympy==1.11.1
TBB==0.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==2.0.0
torch-summary==1.4.5
torch-tb-profiler==0.4.1
torchaudio==2.0.1
torchsummary==1.5.1
torchvision==0.15.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
